adding grpic_1d.hpp; update ampere kernel for 1d grpic
